ai/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
ollama==0.2.1
openai==1.35.7
pandas==2.2.2
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
